+++ uefi/CHANGELOG.md
@@ -17,6 +17,12 @@ details of the deprecated items that were removed in this release.
   deprecated conversion from `uefi::table::boot::ScopedProtocol` has been
   removed.
 - Fixed `boot::open_protocol` to properly handle a null interface pointer.
+- `VariableKey` now has a public `name` field. This `name` field always contains
+  a valid string, so the `VariableKey::name()` method has been deprecated. Since
+  all fields of `VariableKey` are now public, the type can be constructed by
+  users.
+- The `VariableKeys` iterator will now yield an error item if a variable name is
+  not UCS-2.
 
 
 # uefi - 0.32.0 (2024-09-09)